refactor(pool): adopt _strict_eq across prefix-match paths after rebase
The rebased #4222 base hardened equality-based merges with _strict_eq
(python == conflates 0/0.0 and True/1, which hash differently). Move it
to _pool.py so the batch condense functions can share it, and apply it
to the occurrence-keyed prefix matches introduced on this branch:
MessagePoolIndex.match_prefix and both batch prefix loops.
