+++ src/inspect_ai/event/_pool.py
@@ -33,11 +33,12 @@ by object identity / id-bucket equality and hash only genuinely new
 content.
 """
 
+import dataclasses
 import json
 from collections.abc import Iterable, Mapping, Sequence
 from typing import Final, TypeVar, cast
 
-from pydantic import JsonValue
+from pydantic import BaseModel, JsonValue
 from pydantic_core import to_jsonable_python
 
 from inspect_ai._util.hash import mm3_hash
@@ -56,6 +57,55 @@ def materialize_pooled_events(
     materialized = validate_events(list(events))
     materialized = resolve_model_event_inputs(materialized, message_pool)
     return resolve_model_event_calls(materialized, call_pool)
+
+
+def _strict_eq(a: object, b: object) -> bool:
+    """Equality that distinguishes values with different JSON serializations.
+
+    Python ``==`` conflates values the pool hashes distinguish: ``0 == 0.0``
+    and ``True == 1``, but ``json.dumps`` emits different bytes for each, so
+    ``_msg_hash``/``_call_hash`` differ. An ``==``-based merge of such values
+    would reuse a pool entry whose stored bytes round-trip to the *other*
+    value — silent data corruption. This comparison requires matching types
+    for scalars (recursing into models, dataclasses, dicts, and lists), so a
+    merge implies identical serialization.
+
+    Lives here (not ``_pool_index``) because both the batch condense
+    functions below and the per-event indices need it, and ``_pool_index``
+    already imports from this module.
+    """
+    if a is b:
+        return True
+    ta, tb = type(a), type(b)
+    if ta is not tb:
+        return False
+    if isinstance(a, BaseModel):
+        return _strict_eq(a.__dict__, b.__dict__)  # type: ignore[attr-defined]
+    if dataclasses.is_dataclass(a) and not isinstance(a, type):
+        return _strict_eq(vars(a), vars(b))
+    if ta is dict:
+        assert isinstance(a, dict) and isinstance(b, dict)
+        if len(a) != len(b):
+            return False
+        sentinel = object()
+        for k, v in a.items():
+            other = b.get(k, sentinel)
+            if other is sentinel or not _strict_eq(v, other):
+                return False
+            # dict lookup matches keys by ==, which conflates 0/0.0 and
+            # True/1 on the key axis just like values ({0: x} and {0.0: x}
+            # serialize to different JSON); str keys (the JSON-bound common
+            # case) cannot ==-collide across types, so only non-str keys
+            # need their counterpart's type checked
+            if type(k) is not str and not any(
+                bk == k and type(bk) is type(k) for bk in b
+            ):
+                return False
+        return True
+    if ta is list or ta is tuple:
+        assert isinstance(a, (list, tuple)) and isinstance(b, (list, tuple))
+        return len(a) == len(b) and all(_strict_eq(x, y) for x, y in zip(a, b))
+    return a == b
 
 
 def _msg_hash(msg: ChatMessage) -> str:
@@ -158,13 +208,15 @@ def condense_model_event_inputs(
                 continue
             if event.input:
                 # longest prefix of the previous event's input (identity,
-                # then equality) — the same occurrence-identity mechanism
-                # as condense_model_event_with_indices
+                # then strict equality) — the same occurrence-identity
+                # mechanism as condense_model_event_with_indices
                 raw_indices: list[int] = []
                 for msg, prev_msg, prev_index in zip(
                     event.input, prev_input, prev_indices
                 ):
-                    if msg is prev_msg or msg == prev_msg:
+                    if msg is prev_msg or (
+                        msg == prev_msg and _strict_eq(msg, prev_msg)
+                    ):
                         raw_indices.append(prev_index)
                     else:
                         break
@@ -283,7 +335,10 @@ def condense_model_event_calls(
             if msgs and isinstance(msgs, list):
                 raw_indices: list[int] = []
                 for msg, prev_msg, prev_index in zip(msgs, prev_msgs, prev_indices):
-                    if msg == prev_msg:
+                    # _strict_eq, not ==: a prefix element drifting 0 -> 0.0
+                    # or True -> 1 is python-equal but hashes differently;
+                    # reusing the pool index would round-trip the other value
+                    if _strict_eq(msg, prev_msg):
                         raw_indices.append(prev_index)
                     else:
                         break
